requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
streamlit==1.43.2
pydantic==2.10.6
elasticsearch==7.17.12
pandas==2.2.3
google-cloud-storage==2.19.0
google-auth==2.38.0
requests==2.32.3
python-multipart==0.0.20